+++ packages/docs/src/components/head.js
@@ -12,14 +12,14 @@ export default props => {
     .join(' — ')
 
   return (
-    <Helmet>
+    <Helmet htmlAttributes={{ lang: 'en-US' }}>
       <title>{title}</title>
       <meta name="description" content={pkg.description} />
       <link rel="icon" type="image/png" href="/icon.png" />
       <link rel="apple-touch-icon" type="image/png" href="/icon.png" />
       <meta name="twitter:card" content="summary" />
       <meta name="twitter:site" content="@jxnblk" />
-      <meta name="twitter:image" content="https://theme-ui.com/icon.png" />
+      <meta name="twitter:image" content="https://theme-ui.com/card.png" />
       <meta name="twitter:title" content={title} />
       <meta name="twitter:description" content={pkg.description} />
     </Helmet>