+++ packages/docs/src/gatsby-plugin-theme-ui/index.js
@@ -1,7 +1,3 @@
-// other themes can be imported to use as an extension
-// import baseTheme from 'demo-theme/src/gatsby-theme-ui'
-// console.log('base theme', typeof baseTheme, baseTheme)
-
 const heading = {
   fontFamily: 'heading',
   fontWeight: 'heading',
@@ -19,10 +15,6 @@ export default {
     highlight: '#ffffcc',
     gray: '#777',
     purple: '#609',
-    header: {
-      text: 'inherit',
-      background: 'inherit',
-    },
     modes: {
       dark: {
         text: '#fff',
@@ -70,16 +62,13 @@ export default {
   },
   textStyles: {
     heading,
+    display: {
+      variant: 'textStyles.heading',
+      fontSize: [5, 6, 7],
+      mt: 3,
+    },
   },
   styles: {
-    Header: {
-      color: 'header.text',
-      bg: 'header.background',
-    },
-    Footer: {
-      color: 'header.text',
-      bg: 'header.background',
-    },
     Container: {
       p: 3,
       maxWidth: 1024,
@@ -90,8 +79,7 @@ export default {
       fontWeight: 'body',
     },
     h1: {
-      variant: 'textStyles.heading',
-      fontSize: 6,
+      variant: 'textStyles.display',
     },
     h2: {
       variant: 'textStyles.heading',