+++ packages/gatsby-plugin-theme-ui/gatsby-browser.js
@@ -1,1 +1,1 @@
-exports.wrapRootElement = require('./root')
+export { wrapRootElement } from './src/provider'