+++ packages/gatsby-plugin-theme-ui/gatsby-ssr.js
@@ -1,20 +1,18 @@
-const { jsx } = require('theme-ui')
+import { jsx } from 'theme-ui'
 
-exports.wrapRootElement = require('./root')
+export { wrapRootElement } from './src/provider'
 
-// prevent color mode flash
 const noflash = `
 (function() {
   try {
     var mode = localStorage.getItem('theme-ui-color-mode');
     if (!mode) return
     document.body.classList.add('theme-ui-' + mode);
-  } catch (e) {
-  }
+  } catch (e) {}
 })();
 `
 
-exports.onRenderBody = ({ setPreBodyComponents }) => {
+export const onRenderBody = ({ setPreBodyComponents }) => {
   const script = jsx('script', {
     dangerouslySetInnerHTML: {
       __html: noflash,